Aggregated multiple test files using CuPy into single test_device_copy_cupy
